Moved Log to a separate project/assembly for Unity support purposes
Building the logging system into its own .dll is necessary for the Unity logging system to also build into its own .dll, which allows the Unity console to skip the logging system in call stacks and jump directly to the true source of a message.

Added `DefaultLogHandlerAttribute` to allow Log to initialize with the correct default logger immediately, instead of after said logger gets initialized and sets itself as default at a later time.