Adding Unity .asmdef to allow Unity projects to work out of the box
Attempts were made to auto-generate this file for new projects, but barriers led to the conclusion that storing this file in the C# core repo is the simplest and least error prone solution. While this does leave a Unity-specific file in this repo, it is harmless and allows new Unity projects to import both submodules and begin working immediately, without requiring manual creation of this .asmdef or any setup steps at all.