UpdateHandle - Add IsSourceUpdating
Add boolean flags to the update sources and UpdateHandle to indicate when they are within their update phase. This allows executing code to take shortcuts if it knows it's currently executing during an appropriate time in the frame.
Ex: If a system needs to conduct work during LateUpdate but doesn't know if it's executing during that phase (or during the UpdateHandle's update call stack) it can check this flag to execute it's work immediately instead of scheduling to wait for the next Late Update.