Log - Add Log.SupressLogging, Log.IsHandlingLog, ILogListener
Log.SupressLogging - Allows logging to handlers to be suppressed.
Log.IsHandlingLog - Indicates when Log is looping through its handlers for a given log message. Used to prevent recursion when a log listener and log handler share the same system
LogListener - Interface for listeners that pick up log messages from a built in system and re-direct them through the logging system