Changed Logging library .NET Framework from 4.8 to 4.7.1
This allows the project to load correctly in VS Code, because Unity generates .csproj files targeting .NET Framework 4.7.1, and the C#/Omnisharp extension is too dumb to use a library targeting 4.8. The change is harmless (the Logging library does not use any 4.8-specific features) but allows the use of VS Code as a Unity IDE without manually editing any .csproj files.