Handle log caller file paths in a platform-independent way
This fixes a specific issue with log file paths - when logs are emitted from the logging DLLs, they retain the file paths and line numbers of the source code at the time of compilation. This means that if the DLLs were last compiled on a Windows machine, and then an app using Anvil is run on Mac, a Windows file path is given to the [CallerFilePath] attribute. Passing this to `Path.GetFileNameWithoutExtension` fails to recognize backslashes as path separators and returns the entire path (without file extension).